% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmatrix.R
\name{h.rsmatrix}
\alias{h.rsmatrix}
\title{wrapper for MCMCpack's rdirichlet func}
\usage{
h.rsmatrix(alpha)
}
\arguments{
\item{alpha}{the alpha matrix to be supplied to the rdirichlet function}
}
\value{
A dirichlet random vector
}
\description{
Generate dirichlet random vectors without specifying 'n' (assumed
    to be 1)
}
