% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2c.R
\name{l2c}
\alias{l2c}
\title{Loadings to Correlation Matrix}
\usage{
l2c(loadings)
}
\arguments{
\item{loadings}{the loading matrix to be utilized.}
}
\value{
A correlation matrix
}
\description{
Easy method to generate correlation matricies from factor
   loadings (either made up or real)
}
\examples{
loadings <- matrix(c(0.6,0.7,0.1,0.1,0.2,
                     0.3,0.1,0.7,0.7,0.1,
                     0,0.1,0.2,0.2,0.8),
                   ncol = 3)
l2c(loadings)


}
