% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasta.R
\name{lasagna_mc}
\alias{lasagna_mc}
\title{Lasagna plot (Updated from bruce's)}
\usage{
lasagna_mc(ob, data)
}
\arguments{
\item{ob}{which row to plot - will change to make more general}

\item{data}{data frame to plot}
}
\value{
A lasagna plot
}
\description{
Generates a good looking plot #pasta
}
