% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{psych_parcer}
\alias{psych_parcer}
\title{parces}
\usage{
psych_parcer(object, correlation)
}
\arguments{
\item{object}{the psych object}

\item{correlation}{a corr matrix}
}
\value{
comparison
}
\description{
retrievs info
}
