% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmatrix.R
\name{rsmatrix}
\alias{rsmatrix}
\title{Generate random strucutred corr plots}
\usage{
rsmatrix(struc)
}
\arguments{
\item{struc}{the 'weights' matrix}
}
\value{
A correlation matrix
}
\description{
Easy method to generate random strucuted correlation matricies
}
\examples{
struc <- matrix(c(5,1,2,
                  4,1,1,
                  7,1,1.5,
                  1,6,1,
                  1,7,1,
                  1,5,1,
                  1,1,9,
                  1,1,9,
                  1,1,5)
                , ncol= 3)
rsmatrix(struc)


}
