% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2c.R
\name{tester.l2c}
\alias{tester.l2c}
\title{l2c helper function}
\usage{
tester.l2c(loadings)
}
\arguments{
\item{loadings}{the loading matrix to be utilized.}
}
\description{
determines if a proper loadings matrix has been provided

tests to make sure a valid factor loadings matrix is provided
}
